quiet compiler warning
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@1172 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ lib/libpgm1.c
@@ -196,7 +196,7 @@ readRpgmRow(FILE * const fileP,
                       errno, strerror(errno));
         else if (rc != bytesPerRow)
             asprintfN(&error, "Error reading row.  Short read of %u bytes "
-                      "instead of %u", rc, bytesPerRow);
+                      "instead of %u", (unsigned)rc, bytesPerRow);
         else {
             error = NULL;
             if (maxval < 256) {
